+++ examples/ldc/ldc2d_steady.py
@@ -68,12 +68,7 @@ pde_disc = psci.discretize(pde, space_npoints=10201, space_method="uniform")
 # Network
 # TODO: remove num_ins and num_outs
 net = psci.network.FCNet(
-    num_ins=2,
-    num_outs=3,
-    num_layers=10,
-    hidden_size=50,
-    dtype="float32",
-    activation='tanh')
+    num_ins=2, num_outs=3, num_layers=10, hidden_size=50, activation='tanh')
 
 # Loss
 loss = psci.loss.L2()
